% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFits.R
\name{colony.fit2}
\alias{colony.fit2}
\title{Internal QFA function called by qfa.fit2. Do not call directly}
\usage{
colony.fit2(
  position,
  bcdata,
  inocguess,
  fixG = TRUE,
  globalOpt = FALSE,
  detectThresh = 0,
  minK = 0,
  logTransform = FALSE,
  AUCLim = 5,
  STP = 10,
  glog = glog,
  modelFit = TRUE,
  checkSlow = TRUE,
  nrate = TRUE,
  TimeFormat = "d",
  Gmp = F,
  lowK = NA,
  upK = NA,
  lowr = NA,
  upr = NA,
  lowg = NA,
  upg = NA,
  lowv = NA,
  upv = NA,
  lowb = NA,
  upb = NA,
  Nrm = "Raw",
  ...
)
}
\description{
Internal QFA function called by qfa.fit2. Do not call directly
}
