% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{colony.info}
\alias{colony.info}
\title{Internal QFA function. Do not call directly}
\usage{
colony.info(position, bcdata)
}
\description{
Extracting meta data of a position from the barcode metadata
}
