% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFits.R
\name{data.fit2}
\alias{data.fit2}
\title{Internal QFA function called by qfa.fit2. Do not call directly}
\usage{
data.fit2(
  tim,
  growth,
  inocguess,
  xybounds,
  inits = list(),
  logTransform = FALSE,
  verbose = FALSE,
  TimeFormat = "d",
  Gmp = F,
  Nrm = "Raw"
)
}
\description{
Internal QFA function called by qfa.fit2. Do not call directly
}
