% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFits.R
\name{de.fit2}
\alias{de.fit2}
\title{Internal QFA function called by qfa.fit2. Do not call directly}
\usage{
de.fit2(
  tim,
  growth,
  inocguess,
  xybounds,
  inits = list(),
  initPop = FALSE,
  widenr = F,
  logTransform = FALSE,
  mxit = 2000,
  Gmp = F,
  Nrm = "Raw"
)
}
\description{
Internal QFA function called by qfa.fit2. Do not call directly
}
