% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertPlateReader.R
\name{GC2QFA}
\alias{GC2QFA}
\title{Converting plate reader growth curve output for QFA usage}
\usage{
GC2QFA(
  GC,
  info,
  rows = 8,
  cols = 12,
  correction = "blank",
  timeFormat = "min",
  ExpName = "Default",
  Inoc.Time = "2019-01-20 03:21:00",
  Library.Name = "Default",
  MasterPlate.Number = 1,
  ScreenID = "SaureusTest",
  Screen.Name = "SauresTest",
  Treatments = 1,
  shiftZero = F
)
}
\arguments{
\item{GC}{data.frame. This contains the growth curve data from a plate reader stored in 'Time' format. Column names should represent well names.
One column needs to be named 'Time' and indicate the timecourse of the growth curve either in minutes, hours or days. No other columns should be present}

\item{info}{data.frame. This contains the metadata of the GC data.frame. Two columns are necessary: One named 'Well' corresponding to the column names
of the GC data.frame and a second named 'Condition' describing the well contents. All other columns are added to the final output of GC2QFA}

\item{rows}{numerical. Indicating number of rows of the growth curve plate. Default set to 8 for 96 well formats}

\item{cols}{numerical. Indicating number of cols of the growth curve plate. Default set to 12 for 96 well formats. Please note that the length of the
info data.frame needs to match rows times cols and that the number of columns of the GC data.frame needs to match rows times cols + 1 (for the time column)}

\item{correction}{String. Indicating the type of correction method used. Either 'blank', 'individual' or 'none'}

\item{timeFormat}{String. Indicating the format of the 'Time' column of the GC data.frame. Either 'min', 'h' or 'd'}

\item{ExpName}{String. Necesarry information for qfa.fit2 to describe the experiment (Barcode). Use this as identifier for each plate/run to ensure
smooth operation of qfa.fit2 and following functions}

\item{Inoc.Time}{String. Inoculation time start in specific format 'YYYY-MM-DD hh:mm:ss'. Necesarry for qfa.fit2 to calculate timecourse}

\item{Library.Name}{String. Necesarry information for qfa.fit2 to describe the experiment}

\item{MasterPlate.Number}{Numerical. Necesarry information for qfa.fit2 to describe the experiment}

\item{ScreenID}{String. Necesarry information for qfa.fit2 to describe the experiment}

\item{Screen.Name}{String. Necesarry information for qfa.fit2 to describe the experiment}

\item{Treatments}{Numerical or String. Necesarry information for qfa.fit2 to describe the experiment}

\item{shiftZero}{logical. Indicating if each wells growth curve should be normalised to start from 0 by subtracting from each well its minimum}
}
\description{
This function transforms the output of plate reader growth curves saved in 'Time' format to a data.frame suitable for calls to qfa.fit2.
Together with the growth curve data (GC) as a data.frame, a second data.frame containing metadata information (info) about the wells of the GC
data.frame. There are specific formatting conditions that need to be met by the input data.frames. There needs to be a column named 'Time'
or 'time' indicating time for the GC data.frame. This column should either be in hours, min or days and this information is specified with
'timeFormat'. The other column names of the GC data.frame should indicate the 'Well' name. Each of these wells need to be represented with one
row in the info data.frame. The info data.frame needs to have a calumn called 'Well' with these matching well names. The second mandatory
column should be named 'Condition' and describes the content of each well. Here, the user can also specify which wells are blanks to be used
as a correction mean. Additionally, the user needs to specify the numbers of rows and cols. The default values are for 96 well formats.
The output data.frame contains aggregated data with one row per well and time combination together with all info data.frame data
and is ready to be used with qfa.fit2. The output data is stored in columns named 'OD_raw' and 'OD_calib'.
}
\details{
Different calibration methods can be used. The default method is 'blank'. A mean of all wells indicated as 'blank' within the
info data.frame column 'Condition' for each measurement timepoint is calculated. This mean is then subtracted from all other wells.
Another options is 'individual': the min value of each well is subtracted from itself. With 'none', no calibration is calculated.
}
