% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{gen_logistic_maxslp}
\alias{gen_logistic_maxslp}
\title{Internal QFA function. Do not call directly}
\usage{
gen_logistic_maxslp(K, r, g, v)
}
\description{
Maximum slope of generalised logistic growth curve (on linear scale)
}
