% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFits.R
\name{growthcurve2}
\alias{growthcurve2}
\title{Internal QFA function called by qfa.fit2. Do not call directly}
\usage{
growthcurve2(
  obsdat,
  iguess,
  fixG = TRUE,
  globalOpt = FALSE,
  detectThresh = 0,
  minK = 0,
  logTransform = FALSE,
  glog = glog,
  checkSlow = TRUE,
  TimeFormat = "d",
  Gmp = F,
  lowK = NA,
  upK = NA,
  lowr = NA,
  upr = NA,
  lowg = NA,
  upg = NA,
  lowv = NA,
  upv = NA,
  lowb = NA,
  upb = NA,
  Nrm = "Raw",
  ...
)
}
\description{
Internal QFA function called by qfa.fit2. Do not call directly
}
