% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFits.R
\name{guess2}
\alias{guess2}
\title{Internal QFA function called by qfa.fit2. Do not call directly}
\usage{
guess2(
  tim,
  growth,
  inocguess,
  xybounds,
  minK = 0.025,
  TimeFormat = "d",
  Gmp = F
)
}
\description{
Internal QFA function called by qfa.fit2. Do not call directly
}
