% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFits.R
\name{loapproxfun}
\alias{loapproxfun}
\title{Internal QFA function. Do not call directly}
\usage{
loapproxfun(t, g, span = 0.2)
}
\description{
Closure returning a loess-based approximating function for a timeseries
}
