% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFits.R
\name{makeBoundsQFA2}
\alias{makeBoundsQFA2}
\title{Internal QFA function called by qfa.fit2. Do not call directly}
\usage{
makeBoundsQFA2(
  inocguess,
  d,
  minK = 0,
  fixG = FALSE,
  globalOpt = FALSE,
  glog = glog,
  TimeFormat = "d",
  Gmp = F,
  lowK = NA,
  upK = NA,
  lowr = NA,
  upr = NA,
  lowg = NA,
  upg = NA,
  lowv = NA,
  upv = NA,
  lowb = NA,
  upb = NA,
  Nrm = "Raw",
  ...
)
}
\description{
Internal QFA function called by qfa.fit2. Do not call directly
}
