% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFits.R
\name{numerical_r2}
\alias{numerical_r2}
\title{Internal QFA function called by qfa.fit2. Do not call directly}
\usage{
numerical_r2(
  obsdat,
  mkPlots = F,
  span = 0.3,
  nBrute = 10000,
  cDiffDelta = 1e-04,
  mlab = ""
)
}
\description{
Internal QFA function called by qfa.fit2. Do not call directly
}
