% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitnessEstimation.R
\name{numericalfitness2}
\alias{numericalfitness2}
\title{Internal QFA function. Do not call directly.}
\usage{
numericalfitness2(obsdat, AUCLim, STP, nrate = T)
}
\description{
This function does numerical fitness estimation by calculating numerical AUC, max slope and time to max slope and others.
}
