% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePlots.R
\name{qfa.AggregateCurves}
\alias{qfa.AggregateCurves}
\title{Plot all data measurement points including model curves grouped by userinput in a ggplot object}
\usage{
qfa.AggregateCurves(
  results,
  GC,
  plotCol = "ORF",
  plotFacet = "ScreenID",
  yname = "Growth",
  legName = NA,
  PlotName = NA,
  markb = F,
  TimeFormat = NA
)
}
\arguments{
\item{results}{data.frame. Usually, the output data.frame of qfa.fit2 to create model curves. It can also be the output data.frame of Read.Colonyzer
or GC2QFA if the data without model curves should be displayed}

\item{GC}{data.frame. The data.frame used as an input for qfa.fit2 to generate the results data.frame. If the results data.frame is the output
of either Read.Colonyzer or GC2QFA, then GC needs to be the same data.frame}

\item{plotCol}{String. The name of the column header which should be used for coloring of the curves}

\item{plotFacet}{String. The name of the column header which should be used for facetting of the plots}

\item{yname}{String. Label of the y-axis}

\item{legName}{String. Label for the color legend}

\item{PlotName}{String. Title for the plot}

\item{markb}{logical. Indicator if the b parameter of the Gompertz model should be marked as a
colored vertical line. Ignored if another model was used to create the results data.frame}

\item{TimeFormat}{String. Either "d" for days or "h" for hours. If set to NA and results contains model-fits, the TimeFormat
specified in the call to qfa.fit2 is used}
}
\value{
This function just plots the ggplot object if not used with an assignment. If it is
called as an assignment, the return is a ggplot object which can further be modified by
the user if desired.
}
\description{
This functions creates a ggplot object with the input data.frames from either Read.Colonyzer
or GC2QFA functions (data.frame GC) and the model fits from a subsequent call to qfa.fit2
(data.frame results) of the same data.frame. The data can also be displayed without the model-fits.
For this, the user inputs the same data.frame for results and GC which was generated by either Read.Colonyzer
or GC2QFA. The qfa.AggregateCurves functions lets the user specify by which factors the plot should be facetted and colored.
These factors need to be present in both results and GC data.frame as a column.
}
\examples{
#qfa.testdata was generated with the call in the Not run section
#(these files are not included in the package)
data(qfa.testdata)
#Strip non-experimental edge cultures
qfa.testdata = qfa.testdata[(qfa.testdata$Row!=1) & (qfa.testdata$Col!=1) & (qfa.testdata$Row!=8) & (qfa.testdata$Col!=12),]
# Define which measure of cell density to use
qfa.testdata$Growth = qfa.testdata$Intensity
GmpFit = qfa.fit2(qfa.testdata, inocguess=NULL, detectThresh=0, globalOpt=F, AUCLim=NA, TimeFormat="h", Model="Gmp")
# Construct fitness measures
GmpFit = makeFitness2(GmpFit, AUCLim=NA, plotFitness="All", filename="Example_Gmp_fitness.pdf")
# Create aggregated curves plot
a = qfa.AggregateCurves(GmpFit, qfa.testdata, plotFacet="ORF", yname="Intesity [AU]", legName="Strain", markb=F)
# Demonstrating subsequent ggplot object manipulations:
a = a + ggplot2::ggtitle("Alternate title")
a

}
\keyword{qfa}
