% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyData.R
\name{qfa.normalize}
\alias{qfa.normalize}
\title{Function to normalize QFA data}
\usage{
qfa.normalize(
  dta,
  NormMode = "All",
  AdjTo = NA,
  AdjInd = NA,
  MaxVal = NA,
  MinVal = NA
)
}
\arguments{
\item{NormMode}{String. This indicates the method used for the normalisation. For explanations, see the text above.
Possible options: All, Plate, Sample, ZeroShift, Adjust, Reference.}

\item{AdjTo}{Numerical. Only necesarry to specify for the Adjust method. Indicates the index to
take the value from to replace at index AdjInd.}

\item{AdjInd}{Numerical. Only necessary to specify for the Adjust method. Indicates the index of the
growth value to be replaced with the value at index AdjTo.}

\item{MaxVal}{Numerical. Only necessary to specify for the Reference method. This
indicates the potential max value that could have been recorded.}

\item{MinVal}{Numerical. Only necessary to specify for the Reference method. This
indicates the potential min value that could have been recorded.}

\item{path}{dta data.frame. This should contain data generated by either GC2QFA or colonyzer.read and a column
named Growth in which the growth data is stored. This column's data will be normalized.}
}
\description{
This function normalizes the input data to be in the range of o to 1.
(Input data.frame preferentially generated by colonyzer.read
or GC2QFA). The target values need to be stored in a column named Growth.
Different methods for normalization are provided. "All", the default methods,
normalizes all data within the data.frame. The method "Plate" normalizes per plate
indicated by a distinct "Barcode" within the data.frame. Method "Sample"
normalizes over each position indicated by Row and COlumn and distinct Barcode.
Method "ZeroShift" subtracts from each position its own minimal value to move the
starting values of the growth curve to zero. Method "Adjust" is useful if at one timepoint
an arbitrarily high value was recorded. This can happen if the light was considerably
higher at one timepoint due to experimental error. If this only happens infrequently,
the user can specify the AdjInd as the position of measurement to be replaced with
the value of AdjTo. Usually AdjTo should be AdjInd+/-1. Use cautiosly.
The last method "Reference" normalizes the given data but does not take the min and max values
of itself but the values MaxVal and MinVal provided by the user. This may be useful if comparing
datasets generated with slightly varying imaging settings.
}
