% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfa.testdata.R
\docType{data}
\name{qfa.testdata}
\alias{qfa.testdata}
\title{A screen of three Staphylococcus aureus strains arranged in a grid}
\format{
data.frame
}
\usage{
data(qfa.testdata)
}
\description{
This is the result of a call of Colonyzer.read of an experiment in which
three different Staphylococcus aureus strains have been grown in a grid.
The three strains are Cowan and JE2 lab strains and a clinical isolate
named CI1149.
}
\examples{
data(qfa.testdata)
#Strip non-experimental edge cultures
qfa.testdata=qfa.testdata[(qfa.testdata$Row!=1) & (qfa.testdata$Col!=1) & (qfa.testdata$Row!=8) & (qfa.testdata$Col!=12),]
# Define which measure of cell density to use
qfa.testdata$Growth = qfa.testdata$Intensity
GmpFit = qfa.fit2(qfa.testdata, inocguess=NULL, detectThresh=0, globalOpt=F, AUCLim=NA, TimeFormat="h", Model="Gmp")
# Construct fitness measures
GmpFit = makeFitness2(GmpFit, AUCLim=NA, plotFitness="All", filename="Example_Gmp_fitness.pdf")
# Create plot
qfa.plot2("Example_Gmp_GrowthCurves.pdf", GmpFit, qfa.testdata, maxt=30)
}
\references{
unpublished
}
\keyword{datasets}
