% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{qfa_check_packages}
\alias{qfa_check_packages}
\title{Internal QFA function. Do not call directly.}
\usage{
qfa_check_packages()
}
\description{
Internal QFA function to check if pachages for plotting are installed.
}
