% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFits.R
\name{sumsq2}
\alias{sumsq2}
\title{Internal QFA function called by qfa.fit2. Do not call directly}
\usage{
sumsq2(K, r, b, growth, tim, logTransform = FALSE, Gmp = F, yShift = 0)
}
\description{
Internal QFA function called by qfa.fit2. Do not call directly
}
